Extract emoji picker as component
