+++ src/index.ts
@@ -1,0 +1,2 @@
+export * from "./components/EmojiPicker";
+export * from "./components/Emoji";
